pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectral_explorer"
version = "0.1.0"
description = "A package for running a text game about exploring generated worlds"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
  { name = "Nicholas Hotelling", email = "nicholashotelling@gmail.com" },
]
license = { text = "MIT" }
dependencies = [
  "openai",
  "pillow",
]
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = []